refactor: CalendarEventRequest를 domain/entities로 이동 (DIP 위반 수정)
domain 레이어 repository 인터페이스가 data 레이어 타입을 직접 import하던 문제 수정
CalendarEventRequest는 외부 의존성 없는 순수 도메인 모델이므로 domain/entities로 이동
