+++ lawding_flutter/lib/domain/usecases/update_calendar_event_usecase.dart
@@ -1,7 +1,7 @@
 import '../core/result.dart';
 import '../entities/calendar_event.dart';
 import '../repositories/calendar_event_repository.dart';
-import '../../data/calendar_event/calendar_event_request.dart';
+import '../entities/calendar_event_request.dart';
 import '../../data/network/network_error.dart';
 
 /// 캘린더 이벤트 수정 UseCase
